+++ code/client/cl_keys.c
@@ -786,6 +786,33 @@ void Key_ClearStates( void )
 }
 
 
+/*
+====================
+Key_KeynumToStringBuf
+====================
+*/
+void Key_KeynumToStringBuf(int keynum, char *buf, int buflen) {
+    Q_strncpyz(buf, Key_KeynumToString(keynum), buflen);
+}
+
+/*
+====================
+Key_GetBindingBuf
+====================
+*/
+void Key_GetBindingBuf(int keynum, char *buf, int buflen) {
+    const char    *value;
+
+    value = Key_GetBinding(keynum);
+
+    if(value) {
+        Q_strncpyz(buf, value, buflen);
+    } else {
+        *buf = 0;
+    }
+}
+
+
 static qint keyCatchers = 0;
 
 /*