+++ code/game/g_public.h
@@ -245,6 +245,7 @@ typedef enum {
   G_TESTPRINTFLOAT,
 
   //engine extensions
+  G_CVAR_SETDESCRIPTION,
   G_TRAP_GETVALUE = COM_TRAP_GETVALUE
 
 } gameImport_t;