+++ code/qcommon/md5.c
@@ -318,6 +318,38 @@ qchar *Com_MD5File( const qchar *fn, qint length, const qchar *prefix, qint pref
 	return final;
 }
 
+qchar *
+Com_MD5Buf(const qchar *data, qint length, const qchar *data2, qint length2)
+{
+  static qchar final_buf[MD5_DIGEST_SIZE * 2 + 1];
+  unsigned qchar digest[MD5_DIGEST_SIZE];
+  unsigned i;
+  MD5_CTX md5;
+
+  MD5Init(&md5);
+
+  if (data && length > 0)
+  {
+    MD5Update(&md5, (unsigned qchar *)data, length);
+  }
+
+  if (data2 && length2 > 0)
+  {
+    MD5Update(&md5, (unsigned qchar *)data2, length2);
+  }
+
+  MD5Final(&md5, digest);
+
+  final_buf[0] = '\0';
+
+  for(i = 0;i < sizeof(digest);i++)
+  {
+    Q_strcat(final_buf, sizeof(final_buf), va("%02X", digest[i] & 0xFF));
+  }
+
+  return final_buf;
+}
+
 //stateless challenges
 static struct MD5Context hmac_ctx_in;
 static struct MD5Context hmac_ctx_out;